Enable rugged as an alternative backend to grit
Grit seems to be no longer maintained, and using it with new versions of Ruby causes encoding errors.

Rugged is in beta right now, but both rugged and grit can co-exist, letting the user of the container specify which backend he/she would like to use.

There is a temporary hack in here that builds the gitlab-grit gem directly from the repo instead of pulling from rubygems. The reason is that there is a conflict in the mime-types dependency between grit and rugged. There is a merged PR that updates the dependency on grit, but the new version has not been pushed to rubygems (see https://github.com/gitlabhq/grit/pull/53). Once the new version is on rubygems, we can simply add the gem to the gemfile instead of building from the repo.

For context on the original issue that caused the move from grit, see this issue: https://github.com/gollum/gollum/issues/843